+++ apps/web/src/app/(auth-pages)/sign-in/page.tsx
@@ -3,7 +3,8 @@ import Link from 'next/link';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 
-import { signInAction } from '@/app/actions';
+import { signInAction } from '@/lib/auth-actions';
+
 import { FormMessage, Message } from '@/features/auth/form-message';
 import { SubmitButton } from '@/features/auth/submit-button';
