+++ apps/web/src/features/auth/header-auth.tsx
@@ -1,6 +1,7 @@
 import Link from 'next/link';
 
-import { signOutAction } from '@/app/actions';
+import { signOutAction } from '@/lib/auth-actions';
+
 import { hasEnvVars } from '@/utils/supabase/check-env-vars';
 import { createClient } from '@/utils/supabase/server';
