+++ apps/web/src/lib/auth-actions.ts
@@ -64,7 +64,7 @@ export const forgotPasswordAction = async (formData: FormData) => {
   }
 
   const { error } = await supabase.auth.resetPasswordForEmail(email, {
-    redirectTo: `${origin}/auth/callback?redirect_to=/protected/reset-password`,
+    redirectTo: `${origin}/auth/callback?redirect_to=/reset-password`,
   });
 
   if (error) {
@@ -90,15 +90,11 @@ export const resetPasswordAction = async (formData: FormData) => {
   const confirmPassword = formData.get('confirmPassword') as string;
 
   if (!password || !confirmPassword) {
-    encodedRedirect(
-      'error',
-      '/protected/reset-password',
-      'Password and confirm password are required',
-    );
+    encodedRedirect('error', '/reset-password', 'Password and confirm password are required');
   }
 
   if (password !== confirmPassword) {
-    encodedRedirect('error', '/protected/reset-password', 'Passwords do not match');
+    encodedRedirect('error', '/reset-password', 'Passwords do not match');
   }
 
   const { error } = await supabase.auth.updateUser({
@@ -106,10 +102,10 @@ export const resetPasswordAction = async (formData: FormData) => {
   });
 
   if (error) {
-    encodedRedirect('error', '/protected/reset-password', 'Password update failed');
+    encodedRedirect('error', '/reset-password', 'Password update failed');
   }
 
-  encodedRedirect('success', '/protected/reset-password', 'Password updated');
+  encodedRedirect('success', '/reset-password', 'Password updated');
 };
 
 export const signOutAction = async () => {
